refactor(model): share clip resize rules
