+++ src/libs/ProjectModel/Utils/ClipResizeUtils.cpp
