+++ src/libs/ProjectModel/Utils/ClipResizeUtils.h
